Adapt GET query arguments syntax in functional tests.
The migration of functional tests to Python3 introduced a wrapper function ftstestclient.py::_adapt_test() meant to adapt the request syntax from Pylons to Flask.
For Python environments installed via "pip", this wrapper doesn't pass properly the query arguments for certain controllers, such as the "/jobs" controller.

This commit changes the way query arguments are passed by passing them directly in the URL instead of relying on a "params" map and the wrapper function.
The URL query arguments approach works on both the RPM and "pip" version