client: add support for --strict-copy flag
Motivation:

There are some circumstances in which it is impossible for FTS to verify
that a transfer was successful: any attempt to do so will fail, which
results in the overall transfer failing.

Rather than failing transfers, it may be acceptable to rely on the
result of the COPY.  If this reports success then consider the transfer
successful.

Support for this already exists in FTS.  It is only the REST client that
is missing a flag.

Modification:

Add the `strict_copy` param with a default value of `False`.  Add a
command-line argument that switches on the behaviour.

Result:

Someone using the REST client is now able to tell FTS that files are to
be copied without FTS attempting to verify the success.