CI: Avoid running empty "merge request" pipelines
Explanation:

When creating a merge request, each commit on the source branch would generate two pipelines: one for the commit on the source branch and another one associated to the merge request. We want to keep the source branch pipeline but do not want the merge request pipeline (we don't run any special "Merge Request" pipeline)

The "Merge Request" pipeline would be empty, having only the "rpms" job. Turns out the following rule triggered it:
```
if: '$CI_COMMIT_REF_NAME != "develop" && $CI_COMMIT_TAG == null'
when: manual
```

The "Merge Request" pipeline fits the criteria: not on "develop" branch and also not a tag.

This change adds an additional constraint, mentioning the source of the pipeline must not be a merge request event.