CI: Use MySQL service instead of MariaDB to run the functional tests
This change was triggered when the Gitlab shared runners infrastructure moved from docker runners to K8s runners. With this change the MariaDB service was not able to start.
Since the MariaDB image being used before was very old (+5 years since the last update) and since FTS runs MySQL in production, these pipelines were changed to use MySQL from now on

In order to be able to connect using the "root" password to the MySQL database, the "mysql_native_password" mode needed to be turned on.
Our CC7 test jobs use mariadb-5.5.68 as the client to connect to the database. This version of MariaDB does not support the new default encryption mode of MySQL 8.X. Therefore, when we start the database we make sure it is compatible with old clients

Note: When running the functional tests, give some time for the MySQL service to start. The MySQL service is heavy and running it on a shared K8s infrastructure requires some time to start