FTS-1959: Implement issuer look-up function to retrieve a registered issuer associated with an access token.
Explanation:
Up until now, the access token's "iss" field was used directly to retrieve the issuer from the configured issuers map. The issuers are registered in the map according to the ".well-known/openid-configuration > issuer" value. There is no guarantee the two fields will match.

To address this, an issuer look-up function was introduced. The function takes the "iss" field from the access token and will look up the issuer by trying both variants, with and without the trailing slash.

This issue was first discovered when integrating with the SKA IAM